refactor(cli): consolidate service-error printing in print.ts
Drop the printServiceFailure -> printServiceError re-export shim that
errors.ts carried only for backward compatibility. The presentation
helpers (printServiceError, formatTransportError) now live solely in
shared/trpc/print.ts under a single name; the command modules that used
them import from there directly instead of routing through
agent/commands/errors.ts. errors.ts keeps only its resolve-error
helpers and consumes printServiceError like everyone else.

No behavior change.
