+++ packages/cli/src/modules/agent/commands/create.ts
@@ -11,7 +11,10 @@ import { parseTimeout } from "../../shared/parse-timeout.js";
 import type { AgentService } from "../services/agent-service.js";
 import { fetchOrFallback } from "../services/fetch-or-fallback.js";
 import { waitForRunning } from "../services/wait-for-state.js";
-import { formatTransportError, printServiceError } from "./errors.js";
+import {
+  formatTransportError,
+  printServiceError,
+} from "../../shared/trpc/print.js";
 import { parseEnvFlag, validateAgentName } from "./create-helpers.js";
 import {
   EXIT_BELOW_FLOOR,
