+++ packages/cli/src/modules/egress/commands/trusted-hosts.ts
@@ -1,5 +1,5 @@
 import { Command } from "commander";
-import { printServiceError } from "../../agent/commands/errors.js";
+import { printServiceError } from "../../shared/trpc/print.js";
 import type { CompatService, ConfigService } from "../../cli/index.js";
 import {
   EXIT_BELOW_FLOOR,
