+++ packages/cli/src/modules/schedule/commands/update.ts
@@ -6,7 +6,7 @@ import {
   detectTimezone,
   hasVisibleOccurrence,
 } from "api-server-api";
-import { printServiceError } from "../../agent/commands/errors.js";
+import { printServiceError } from "../../shared/trpc/print.js";
 import type { CompatService, ConfigService } from "../../cli/index.js";
 import {
   EXIT_BELOW_FLOOR,
